requirements.txt
google-cloud-storage>=2.10.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
python-dotenv
webdriver-manager
functions-framework>=3.4.0
pandas
scikit-learn
umap-learn
matplotlib
hdbscan
google-cloud-bigquery
db-dtypes
pandas
python-dotenv
langdetect
pyarrow